# backend/src/main/resources/static/aas-registry-openapi2.yaml
...
openapi: 3.0.3
info:
  title: DotAAS Part 2 | HTTP/REST | Asset Administration Shell Registry Service Specification
  description: "The Full Profile of the Asset Administration Shell Registry Service Specification as part of the Specification of the Asset Administration Shell: Part 2. Publisher: Industrial Digital Twin Association (IDTA) April 2023"
  contact:
    name: Industrial Digital Twin Association (IDTA)
    email: info@idtwin.org
  license:
    name: CC BY 4.0
    url: https://creativecommons.org/licenses/by/4.0/
  version: V3.0_SSP-001
servers:
- url: "{protocol}://{host_name}:{port}/api/{version_prefix}"
  variables:
    protocol:
      description: Allows access through http and https (recommended)
      default: https
      enum:
      - http
      - https
    host_name:
      description: Hostname of server hosting the api
      default: admin-shell.io
    port:
      description: "80 is default for http, 443 for https"
      default: "443"
      enum:
      - "80"
      - "443"
    version_prefix:
      default: v3.0
      enum:
      - v3.0
paths:
  /shell-descriptors:
    get:
      tags:
      - Asset Administration Shell Registry API
      summary: Returns all Asset Administration Shell Descriptors
      operationId: GetAllAssetAdministrationShellDescriptors
      parameters:
      - name: limit
        in: query
        description: The maximum number of elements in the response array
        required: false
        schema:
          minimum: 1
          type: integer
      - name: cursor
        in: query
        description: A server-generated identifier retrieved from pagingMetadata that specifies from which position the result listing should continue
        required: false
        schema:
          type: string
      - name: assetKind
        in: query
        description: The Asset's kind (Instance or Type)
        required: false
        schema:
          $ref: '#/components/schemas/AssetKind'
      - name: assetType
        in: query
        description: The Asset's type (UTF8-BASE64-URL-encoded)
        required: false
        schema:
          maxLength: 2000
          minLength: 1
          #     pattern: "^[\\x09\\x0A\\x0D\\x20-\\uD7FF\\uE000-\\uFFFD\\U00010000-\\U0010FFFF]*$"
          type: string
      responses:
        "200":
          description: Requested Asset Administration Shell Descriptors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAssetAdministrationShellDescriptorsResult'
        "400":
          description: "Bad Request, e.g. the request parameters of the format of the request body is wrong."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        default:
          description: Default error handling for unmentioned status codes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
      x-semanticIds:
      - https://admin-shell.io/aas/API/GetAllAssetAdministrationShellDescriptors/3/0
      - https://admin-shell.io/aas/API/GetAllAssetAdministrationShellDescriptorsByAssetKind/3/0
      - https://admin-shell.io/aas/API/GetAllAssetAdministrationShellDescriptorsByAssetType/3/0
    post:
      tags:
      - Asset Administration Shell Registry API
      summary: "Creates a new Asset Administration Shell Descriptor, i.e. registers an AAS"
      operationId: PostAssetAdministrationShellDescriptor
      requestBody:
        description: Asset Administration Shell Descriptor object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetAdministrationShellDescriptor'
        required: false
      responses:
        "201":
          description: Asset Administration Shell Descriptor created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetAdministrationShellDescriptor'
        "400":
          description: "Bad Request, e.g. the request parameters of the format of the request body is wrong."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        "409":
          description: "Conflict, a resource which shall be created exists already. Might be thrown if a Submodel or SubmodelElement with the same ShortId is contained in a POST request."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        default:
          description: Default error handling for unmentioned status codes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
      x-semanticIds:
      - https://admin-shell.io/aas/API/PostAssetAdministrationShellDescriptor/3/0
  /shell-descriptors/{aasIdentifier}:
    get:
      tags:
      - Asset Administration Shell Registry API
      summary: Returns a specific Asset Administration Shell Descriptor
      operationId: GetAssetAdministrationShellDescriptorById
      parameters:
      - name: aasIdentifier
        in: path
        description: The Asset Administration Shell’s unique id (UTF8-BASE64-URL-encoded)
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: byte
      responses:
        "200":
          description: Requested Asset Administration Shell Descriptor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetAdministrationShellDescriptor'
        "400":
          description: "Bad Request, e.g. the request parameters of the format of the request body is wrong."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        default:
          description: Default error handling for unmentioned status codes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
      x-semanticIds:
      - https://admin-shell.io/aas/API/GetAssetAdministrationShellDescriptorById/3/0
    put:
      tags:
      - Asset Administration Shell Registry API
      summary: Updates an existing Asset Administration Shell Descriptor
      operationId: PutAssetAdministrationShellDescriptorById
      parameters:
      - name: aasIdentifier
        in: path
        description: The Asset Administration Shell’s unique id (UTF8-BASE64-URL-encoded)
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: byte
      requestBody:
        description: Asset Administration Shell Descriptor object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetAdministrationShellDescriptor'
        required: true
      responses:
        "204":
          description: Asset Administration Shell Descriptor updated successfully
        "400":
          description: "Bad Request, e.g. the request parameters of the format of the request body is wrong."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        default:
          description: Default error handling for unmentioned status codes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
      x-semanticIds:
      - https://admin-shell.io/aas/API/PutAssetAdministrationShellDescriptorById/3/0
    delete:
      tags:
      - Asset Administration Shell Registry API
      summary: "Deletes an Asset Administration Shell Descriptor, i.e. de-registers an AAS"
      operationId: DeleteAssetAdministrationShellDescriptorById
      parameters:
      - name: aasIdentifier
        in: path
        description: The Asset Administration Shell’s unique id (UTF8-BASE64-URL-encoded)
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: byte
      responses:
        "204":
          description: Asset Administration Shell Descriptor deleted successfully
        "400":
          description: "Bad Request, e.g. the request parameters of the format of the request body is wrong."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        default:
          description: Default error handling for unmentioned status codes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
      x-semanticIds:
      - https://admin-shell.io/aas/API/DeleteAssetAdministrationShellDescriptorById/3/0
  /shell-descriptors/{aasIdentifier}/submodel-descriptors:
    get:
      tags:
      - Asset Administration Shell Registry API
      summary: Returns all Submodel Descriptors
      operationId: GetAllSubmodelDescriptorsThroughSuperpath
      parameters:
      - name: aasIdentifier
        in: path
        description: The Asset Administration Shell’s unique id (UTF8-BASE64-URL-encoded)
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: byte
      - name: limit
        in: query
        description: The maximum number of elements in the response array
        schema:
          minimum: 1
          type: integer
      - name: cursor
        in: query
        description: A server-generated identifier retrieved from pagingMetadata that specifies from which position the result listing should continue
        schema:
          type: string
      responses:
        "200":
          description: Requested Submodel Descriptors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetSubmodelDescriptorsResult'
        "400":
          description: "Bad Request, e.g. the request parameters of the format of the request body is wrong."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        default:
          description: Default error handling for unmentioned status codes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
      x-semanticIds:
      - https://admin-shell.io/aas/API/GetAllSubmodelDescriptors/3/0
    post:
      tags:
      - Asset Administration Shell Registry API
      summary: "Creates a new Submodel Descriptor, i.e. registers a submodel"
      operationId: PostSubmodelDescriptor-ThroughSuperpath
      parameters:
      - name: aasIdentifier
        in: path
        description: The Asset Administration Shell’s unique id (UTF8-BASE64-URL-encoded)
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: byte
      requestBody:
        description: Submodel Descriptor object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmodelDescriptor'
      responses:
        "201":
          description: Submodel Descriptor created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmodelDescriptor'
        "400":
          description: "Bad Request, e.g. the request parameters of the format of the request body is wrong."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        "409":
          description: "Conflict, a resource which shall be created exists already. Might be thrown if a Submodel or SubmodelElement with the same ShortId is contained in a POST request."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        default:
          description: Default error handling for unmentioned status codes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
      x-semanticIds:
      - https://admin-shell.io/aas/API/PostSubmodelDescriptor/3/0
  /shell-descriptors/{aasIdentifier}/submodel-descriptors/{submodelIdentifier}:
    get:
      tags:
      - Asset Administration Shell Registry API
      summary: Returns a specific Submodel Descriptor
      operationId: GetSubmodelDescriptorByIdThroughSuperpath
      parameters:
      - name: aasIdentifier
        in: path
        description: The Asset Administration Shell’s unique id (UTF8-BASE64-URL-encoded)
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: byte
      - name: submodelIdentifier
        in: path
        description: The Submodel’s unique id (UTF8-BASE64-URL-encoded)
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: byte
      responses:
        "200":
          description: Requested Submodel Descriptor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmodelDescriptor'
        "400":
          description: "Bad Request, e.g. the request parameters of the format of the request body is wrong."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        default:
          description: Default error handling for unmentioned status codes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
      x-semanticIds:
      - https://admin-shell.io/aas/API/GetSubmodelDescriptorById/3/0
    put:
      tags:
      - Asset Administration Shell Registry API
      summary: Updates an existing Submodel Descriptor
      operationId: PutSubmodelDescriptorByIdThroughSuperpath
      parameters:
      - name: aasIdentifier
        in: path
        description: The Asset Administration Shell’s unique id (UTF8-BASE64-URL-encoded)
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: byte
      - name: submodelIdentifier
        in: path
        description: The Submodel’s unique id (UTF8-BASE64-URL-encoded)
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: byte
      requestBody:
        description: Submodel Descriptor object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmodelDescriptor'
      responses:
        "204":
          description: Submodel Descriptor updated successfully
        "400":
          description: "Bad Request, e.g. the request parameters of the format of the request body is wrong."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        default:
          description: Default error handling for unmentioned status codes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
      x-semanticIds:
      - https://admin-shell.io/aas/API/PutSubmodelDescriptorById/3/0
    delete:
      tags:
      - Asset Administration Shell Registry API
      summary: "Deletes a Submodel Descriptor, i.e. de-registers a submodel"
      operationId: DeleteSubmodelDescriptorByIdThroughSuperpath
      parameters:
      - name: aasIdentifier
        in: path
        description: The Asset Administration Shell’s unique id (UTF8-BASE64-URL-encoded)
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: byte
      - name: submodelIdentifier
        in: path
        description: The Submodel’s unique id (UTF8-BASE64-URL-encoded)
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: byte
      responses:
        "204":
          description: Submodel Descriptor deleted successfully
        "400":
          description: "Bad Request, e.g. the request parameters of the format of the request body is wrong."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        default:
          description: Default error handling for unmentioned status codes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
      x-semanticIds:
      - https://admin-shell.io/aas/API/DeleteSubmodelDescriptorById/3/0
  /description:
    get:
      tags:
      - Description API
      summary: Returns the self-describing information of a network resource (ServiceDescription)
      operationId: GetDescription
      responses:
        "200":
          description: Requested Description
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceDescription'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        default:
          description: Default error handling for unmentioned status codes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
      x-semanticIds:
      - https://admin-shell.io/aas/API/Descriptor/GetDescription/3/0
  /lookup/shells:
    get:
      tags:
        - Registry and Discovery Interface
      summary: Returns a list of Asset Administration Shell ids based on Asset identifier key-value-pairs. Only the Shell ids are returned when all provided key-value pairs match.
      operationId: getAllAssetAdministrationShellIdsByAssetLink
      parameters:
        - name: assetIds
          in: query
          description: The key-value-pair of an Asset identifier
          required: false
          # The form style defined in the AAS API does not match with the provided example.
          # the "content" is the correct way to accept json encoded query parameters
          # style: form
          # explode: true
          content:
            application/json:
              schema:
                type: array
                maxItems: 10000
                items:
                  $ref: '#/components/schemas/SpecificAssetId'
              examples:
                complete:
                  $ref: '#/components/examples/lookup-shells-by-aas-identifier-query'
      responses:
        "200":
          description: Requested Asset Administration Shell ids
          content:
            application/json:
              schema:
                type: array
                maxItems: 10000
                items:
                  $ref: '#/components/schemas/Identifier'
              examples:
                complete:
                  $ref: '#/components/examples/lookup-shells-by-aas-identifier-response'
      x-semanticIds:
        - https://admin-shell.io/aas/API/GetAllAssetAdministrationShellIdsByAssetLink/1/0/RC02
...
  /lookup/shells/query:
    post:
      tags:
        - Registry and Discovery Interface
      summary: Returns a list of Asset Administration Shell ids based on Asset identifier key-value-pairs.
      operationId: PostQueryAllAssetAdministrationShellIds
      requestBody:
        description: Asset Administration Shell Descriptor object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShellLookup'
        required: true
      responses:
        "200":
          description: Requested Asset Administration Shell ids
          content:
            application/json:
              schema:
                type: array
                maxItems: 10000
                items:
                  $ref: '#/components/schemas/Identifier'
              examples:
                complete:
                  $ref: '#/components/examples/lookup-shells-by-aas-identifier-response'
components:
  schemas:
...
    Identifier:
      type: string
      minLength: 1
      maxLength: 200
...
    AssetKind:
      type: string
      enum:
      - Instance
      - NotApplicable
      - Type
    GetAssetAdministrationShellDescriptorsResult:
      allOf:
      - $ref: '#/components/schemas/PagedResult'
      - type: object
        properties:
          result:
            type: array
            items:
              $ref: '#/components/schemas/AssetAdministrationShellDescriptor'
...
    SpecificAssetId:
      allOf:
      - $ref: '#/components/schemas/HasSemantics'
      - properties:
          name:
            maxLength: 64
            minLength: 1
            #  pattern: "^([\\t\\n\\r -퟿-�]|\\ud800[\\udc00-\\udfff]|[\\ud801-\\udbfe][\\udc00-\\udfff]|\\udbff[\\udc00-\\udfff])*$"
            type: string
          value:
            maxLength: 2000
            minLength: 1
            #  pattern: "^([\\t\\n\\r -퟿-�]|\\ud800[\\udc00-\\udfff]|[\\ud801-\\udbfe][\\udc00-\\udfff]|\\udbff[\\udc00-\\udfff])*$"
            type: string
          externalSubjectId:
            $ref: '#/components/schemas/Reference'
...
    Result:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
...
    GetSubmodelDescriptorsResult:
      allOf:
      - $ref: '#/components/schemas/PagedResult'
      - type: object
        properties:
          result:
            type: array
            items:
              $ref: '#/components/schemas/SubmodelDescriptor'
...